refactor: complete DefaultX core architecture cleanup
Introduce gateway-based command and listener registration to remove direct API
dependencies on Paper implementations. Replace registration handles with
logical lifecycle-safe objects and move platform unregister behavior behind
Paper adapters.

Make services constructor-complete, expose missing API accessors, and refactor
core runtime installation into focused private installer phases without
crowding the runtime package or exposing internal runtime methods.

Add the DefaultX annotation layer for API boundaries, runtime wiring, platform
adapters, managed services, dependencies, thread requirements, lifecycle hooks,
and internal APIs. Keep package-level boundary annotations as defaults and use
class-level annotations only for service/runtime-specific behavior.

+++ src/main/java/org/kerix/karaapi/api/lifecycle/Tickable.java
@@ -3,5 +3,5 @@ package org.kerix.karaapi.api.lifecycle;
 public interface Tickable {
     void tick();
 
-    int tickInterval();
+    long tickInterval();
 }
